Modify CME registration options & remove "other" option
Changed two affiliation names. Removed option to select "other"
as an affiliation, and removed validation for the free response field.
Modified js in the cme_register.html file to
omit scrolling animation that reveals a text field when "other" is
selected as affiliation. Since users can no longer submit "other" as an affiliation with a
free-response value, we remove validation for the other_affiliation
post variable.